Move PlayerTechBonuses to DataStructures and fix naming
